Agents/requirements.txt
# Core dependencies from your existing code
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
inference-sdk>=0.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Firebase integration
firebase-admin>=6.2.0

# Flask API
Flask>=2.3.0
Flask-CORS>=4.0.0

# Additional utilities
uuid>=1.30
requests>=2.31.0